clarifications
Clarified what will and won't be in the initial interactive course, per our discussion on Friday. Removed some of the specific topics in psychometrics; it might be a simple thing to add things such as differential item functioning, so not specifically identifying that and other topics as things that won't be covered (although if it's not easy to add them, we won't cover them). Definitely won't be doing multidimensional IRT, however, so identified that as a topic not covered. Also updated the wording a bit.